refactor(idl): extract schema types into standalone crate
Move the IDL type definitions (Idl, IdlInstruction, IdlAccountItem,
IdlType, etc.) from `idl/src/types.rs` into the new `quasar-schema`
crate so derive, IDL parsing, and future client codegen can share one
canonical contract.

Also extract `known_address_for_type()` — the sysvar/program address
lookup — from inline match arms in the IDL parser into quasar-schema
where it's reusable.
