+++ idl/Cargo.toml
@@ -20,3 +20,4 @@ proc-macro2 = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 toml = "0.8"
+quasar-schema = { path = "../schema" }
